move config classes to modeling file
