+++ tensorrt_llm/_torch/auto_deploy/models/eagle.py
@@ -22,60 +22,16 @@ Eagle drafter implementations.
 """
 
 from contextlib import nullcontext
-from dataclasses import dataclass
-from typing import Any, Dict, Type
+from typing import Dict
 
 import torch.nn as nn
 from accelerate import init_empty_weights
 from torch._prims_common import DeviceLikeType
-from transformers import PretrainedConfig, PreTrainedModel
 
-from ..utils._config import deep_merge_dicts
 from ..utils.logger import ad_logger
-from .custom.modeling_eagle import Eagle3DrafterForCausalLM
+from .custom.modeling_eagle import Eagle3DrafterForCausalLM, EagleConfig
 from .factory import ModelFactoryRegistry
 from .hf import AutoModelForCausalLMFactory
-
-
-@dataclass
-class EagleConfigInfo:
-    """Model-specific configuration for Eagle drafters. This class is used to store Eagle-specific information that
-    we use to build the Eagle drafter model - e.g. whether to load the embedding and lm head from the target model.
-    """
-
-    config_class: Type[PreTrainedModel]
-    eagle_config_defaults: dict = None
-
-    def __post_init__(self):
-        if self.eagle_config_defaults is None:
-            self.eagle_config_defaults = {}
-
-
-class EagleConfig(PretrainedConfig):
-    """Config for Eagle3 drafter models.
-
-    Extends PretrainedConfig with Eagle-specific parameters while preserving
-    all base model config values.
-
-    Args:
-        config: Base config for the draft model from its config.json.
-        draft_model_defaults: EagleConfigInfo containing model-specific defaults.
-            These defaults can be overridden by values in config.json.
-    """
-
-    def __init__(self, config: PretrainedConfig, draft_model_defaults: Dict[str, Any]):
-        config_dict = config.to_dict()
-
-        # Log when config overrides a default
-        for key, value in draft_model_defaults.items():
-            if key in config_dict and config_dict[key] != value:
-                ad_logger.info(
-                    f"EagleConfig: config has '{key}={config_dict[key]}', "
-                    f"overriding default '{value}'"
-                )
-
-        merged = deep_merge_dicts(draft_model_defaults, config_dict)
-        super().__init__(**merged)
 
 
 @ModelFactoryRegistry.register("EagleDrafter")
@@ -91,18 +47,8 @@ class EagleDrafterFactory(AutoModelForCausalLMFactory):
     (e.g., "llama") along with Eagle-specific fields like draft_vocab_size.
     """
 
-    # Map config model_type -> EagleConfigInfo (drafter class + model-specific kwargs)
-    # TODO: Possibly we can infer some of these kwargs from the drafter model config, e.g.
-    # load_embedding_from_target can be inferred by the state_dict and/or draft vocab size.
-    _drafter_mapping: Dict[str, EagleConfigInfo] = {
-        "llama": EagleConfigInfo(
-            config_class=Eagle3DrafterForCausalLM,
-            eagle_config_defaults={
-                "load_embedding_from_target": True,
-                "load_lm_head_from_target": False,
-                "num_capture_layers": 3,
-            },
-        ),
+    _drafter_classes: Dict[str, type] = {
+        "llama": Eagle3DrafterForCausalLM,
     }
 
     def _build_model(self, device: DeviceLikeType) -> nn.Module:
@@ -110,20 +56,19 @@ class EagleDrafterFactory(AutoModelForCausalLMFactory):
 
         # Select the appropriate drafter class and config based on the base model type
         model_type = model_config.model_type
-        if model_type not in self._drafter_mapping:
+        if model_type not in self._drafter_classes:
             raise ValueError(
                 f"Unsupported model_type '{model_type}' for Eagle drafter. "
-                f"Supported types: {list(self._drafter_mapping.keys())}"
+                f"Supported types: {list(self._drafter_classes.keys())}"
             )
-        config_info = self._drafter_mapping[model_type]
-        drafter_cls = config_info.config_class
+        drafter_cls = self._drafter_classes[model_type]
         ad_logger.info(
             f"EagleDrafterFactory: model_type='{model_type}' -> drafter_cls={drafter_cls.__name__}"
         )
 
-        # Convert base config to Eagle3Config, preserving existing values
-        # and applying model-specific kwargs from EagleConfigInfo
-        model_config = EagleConfig(model_config, config_info.eagle_config_defaults)
+        # Convert base config to EagleConfig, preserving existing values
+        # and applying model-specific defaults based on model_type
+        model_config = EagleConfig(model_config, model_type)
 
         # Build the model (same pattern as parent's _build_model)
         with (init_empty_weights if device == "meta" else nullcontext)():
